Refactor FreePage to take unsafe.Pointer instead of uintptr
I chose unsafe.Pointer instead of Page as its more work to cast a
pointer back into a slice and FreePage does not use the slice features.
Also for structure that are created from pages the knowledge is
implicit, that it takes an entire page. Also I already have pointers.
it's also easier to cast pointers to unsafe.Pointer instead of casting
it to uintptr. Also go complains less.

+++ kernel/mm/memspace.go
@@ -181,5 +181,5 @@ func (m *MemSpace) FreeAllPages() {
 		tableEntry.UnsetPresent()
 		FreePage(tableEntry.GetPhysicalAddress())
 	}
-	FreePage(uintptr(unsafe.Pointer(m.PageDirectory)))
+	FreePage(unsafe.Pointer(m.PageDirectory))
 }
